# tmpswagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HandySales API - Principal",
    "description": "API REST principal para el sistema HandySales CRM - Gestión de clientes, productos, usuarios y ventas",
    "termsOfService": "https://handysales.com/terms",
    "contact": {
      "name": "HandySales Support",
      "url": "https://handysales.com/support",
      "email": "support@handysales.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0.0"
  },
  "paths": {
...
    "/api/company/billing": {
      "delete": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Eliminar datos de facturación",
        "operationId": "DeleteBillingData",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Obtener datos de facturación",
        "operationId": "GetBillingData",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosFacturacionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Crear datos de facturación",
        "operationId": "CreateBillingData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatosFacturacionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosFacturacionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Actualizar datos de facturación",
        "operationId": "UpdateBillingData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatosFacturacionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosFacturacionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{id}": {
      "delete": {
        "tags": [
          "Company Management"
        ],
        "summary": "Eliminar empresa",
        "description": "Solo accesible para SUPER_ADMIN",
        "operationId": "DeleteCompany",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Company Management"
        ],
        "summary": "Obtener empresa por ID",
        "operationId": "GetCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Company Management"
        ],
        "summary": "Actualizar empresa",
        "operationId": "UpdateCompany",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/company/settings": {
      "get": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Obtener configuración de la empresa",
        "operationId": "GetCompanySettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Actualizar configuración de la empresa",
        "operationId": "UpdateCompanySettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanySettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySettingsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies": {
      "get": {
        "tags": [
          "Company Management"
        ],
        "summary": "Obtener todas las empresas",
        "description": "Solo accesible para SUPER_ADMIN",
        "operationId": "GetAllCompanies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Company Management"
        ],
        "summary": "Crear nueva empresa",
        "description": "Solo accesible para SUPER_ADMIN",
        "operationId": "CreateCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/company/upload-logo": {
      "post": {
        "tags": [
          "Company Settings"
        ],
        "summary": "Subir logo de la empresa",
        "operationId": "UploadCompanyLogo",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadLogoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/impersonation/end": {
      "post": {
        "tags": [
          "Impersonation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndImpersonationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EndImpersonationRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "companyLogo": {
            "type": "string",
            "nullable": true
          },
          "companyPrimaryColor": {
            "type": "string",
            "nullable": true
          },
          "companySecondaryColor": {
            "type": "string",
            "nullable": true
          },
          "companyDescription": {
            "type": "string",
            "nullable": true
          },
          "contactEmail": {
            "type": "string",
            "nullable": true
          },
          "contactPhone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "taxId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateCompanySettingsRequest": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "companyPrimaryColor": {
            "type": "string",
            "nullable": true
          },
          "companySecondaryColor": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UploadLogoResponse": {
        "type": "object",
        "properties": {
          "logoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token JWT en el formato: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}